Move common init/repr into BaseModel
