+++ src/thevenin/_solutions.py
@@ -94,7 +94,7 @@ class BaseSolution(IDAResult):
 
         return readable
 
-    def plot(self, x: str, y: str, show_plot: bool = True, **kwargs) -> None:
+    def plot(self, x: str, y: str, **kwargs) -> None:
         """
         Plot any two variables in 'vars' against each other.
 
@@ -104,10 +104,6 @@ class BaseSolution(IDAResult):
             A variable key in 'vars' to be used for the x-axis.
         y : str
             A variable key in 'vars' to be used for the y-axis.
-        show_plot : bool, optional
-            For non-interactive environments only. When True (default) this
-            registers `plt.show()` to run at the end of the program. If False,
-            you must call `plt.show()` manually.
         **kwargs : dict, optional
             Keyword arguments to pass through to `plt.plot()`. For more info
             please refer to documentation for `maplotlib.pyplot.plot()`.
@@ -136,7 +132,7 @@ class BaseSolution(IDAResult):
         plt.xlabel(xlabel)
         plt.ylabel(ylabel)
 
-        if show_plot and not plt.isinteractive():
+        if not plt.isinteractive():
             ExitHandler.register_atexit(plt.show)
 
     def _fill_vars(self) -> None:
